+++ NEWS.md
@@ -1,5 +1,8 @@
 # purrr 0.2.2.9000
 
+* `map2()` and `pmap()` (and all output type variants) now return a output
+  of length 0 if any input is of length 0.
+
 * `map()`, `map2()`, and `pmap()` (as well as all output type variants)
   treat NULL the same way as an empty vector (#199).